Change progress value to Double

+++ android/src/main/kotlin/vn/hunghd/flutterdownloader/DownloadWorker.kt
@@ -400,7 +400,7 @@ class DownloadWorker(context: Context, params: WorkerParameters) :
                 // using isStopped to monitor canceling task
                 while (inputStream.read(buffer).also { bytesRead = it } != -1 && !isStopped) {
                     count += bytesRead.toLong()
-                    val progress = (count * 100 / (contentLength + downloadedBytes)).toLong()
+                    val progress = (count * 100 / (contentLength + downloadedBytes)).toDouble()
                     outputStream?.write(buffer, 0, bytesRead)
                     if ((lastProgress == 0 || progress > lastProgress + step || progress == 100) &&
                         progress != lastProgress
@@ -622,7 +622,7 @@ class DownloadWorker(context: Context, params: WorkerParameters) :
         context: Context,
         title: String?,
         status: DownloadStatus,
-        progress: Long,
+        progress: Double,
         intent: PendingIntent?,
         finalize: Boolean
     ) {
@@ -711,7 +711,7 @@ class DownloadWorker(context: Context, params: WorkerParameters) :
         }
     }
 
-    private fun sendUpdateProcessEvent(status: DownloadStatus, progress: Long) {
+    private fun sendUpdateProcessEvent(status: DownloadStatus, progress: Double) {
         val args: MutableList<Any> = ArrayList()
         val callbackHandle: Long = inputData.getLong(ARG_CALLBACK_HANDLE, 0)
         args.add(callbackHandle)
